src/Projects/requirements.txt
Flask
mysql-connector-python
pandas
